feat: block user

+++ src/services/updateSettingPostEndMatch.ts
@@ -4,7 +4,7 @@ import { enums } from '../utils';
 
 const updateSettingPostEndMatch = async (userId: number, currentQueue: number) => {
 	const query = `UPDATE setting SET current_queue=$2, avg_match_time=
-        (SELECT ROUND(SUM(score) * 1.0 / COUNT(*), 2) AS avg_match_time FROM match WHERE (user_id_1=$1 OR user_id_2=$1) AND ended=true AND deleted_at IS NULL), 
+        (SELECT ROUND(((SUM(score) * 1.0 / COUNT(*))::numeric), 2) AS avg_match_time FROM match WHERE (user_id_1=$1 OR user_id_2=$1) AND ended=true AND deleted_at IS NULL), 
         is_matched=false WHERE user_id=$1`;
     const params = [userId, currentQueue];
     let res: QueryResult | null = null;
